tests/test_authentication.py
"""
Tests for authentication and session management.
"""

import logging
import os
from unittest.mock import Mock, patch

import pytest

# Try to import Snowflake modules, but handle gracefully if not available
try:
    from snowflake.snowpark import Session
    from snowflake.snowpark.context import get_active_session
    from snowflake.snowpark.exceptions import SnowparkSessionException

    SNOWFLAKE_AVAILABLE = True
except ImportError:
    SNOWFLAKE_AVAILABLE = False

    # Create mock classes for testing
    class Session:
        @staticmethod
        def builder():
            return Mock()

    def get_active_session():
        raise ImportError("Snowflake packages not available")


logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class TestAuthentication:
    """Test authentication scenarios."""

    @pytest.mark.skipif(
        not SNOWFLAKE_AVAILABLE, reason="Snowflake packages not available"
    )
    def test_get_active_session_fails_outside_snowflake(
        self,
    ):  # TODO[6f7a8b9c]: Add type annotations (see TODO.md)
        """Test that get_active_session() fails when not in Snowflake environment."""
        logger.info("Running test_get_active_session_fails_outside_snowflake")
        try:
            with pytest.raises(SnowparkSessionException):
                logger.info(
                    "Calling get_active_session() expecting SnowparkSessionException"
                )
                get_active_session()
        except Exception as e:
            logger.error(
                f"Exception in test_get_active_session_fails_outside_snowflake: {e}"
            )
            raise

    @patch("tests.test_authentication.get_active_session")
    def test_snowflake_environment_session(
        self, mock_get_active_session
    ):  # TODO[7a8b9c0d]: Add type annotations (see TODO.md)
        """Test successful session creation in Snowflake environment."""
        logger.info("Running test_snowflake_environment_session")
        mock_session = Mock()
        mock_get_active_session.return_value = mock_session
        try:
            session = get_active_session()
            logger.info(f"Session returned: {session}")
            assert session == mock_session
            mock_get_active_session.assert_called_once()
        except Exception as e:
            logger.error(f"Exception in test_snowflake_environment_session: {e}")
            raise

    @patch("snowflake.snowpark.session.Session.SessionBuilder.configs")
    @patch("tests.test_authentication.get_active_session")
    def test_local_environment_session_with_env_vars(
        self, mock_get_active_session, mock_configs
    ):  # TODO[8b9c0d1e]: Add type annotations (see TODO.md)
        """Test session creation with environment variables for local development."""
        logger.info("Running test_local_environment_session_with_env_vars")
        mock_get_active_session.side_effect = SnowparkSessionException(
            "No active session"
        )
        mock_configs.return_value = Mock()
        test_config = {
            "SNOWFLAKE_ACCOUNT": "test-account",
            "SNOWFLAKE_USER": "test-user",
            "SNOWFLAKE_PASSWORD": "test-password",
            "SNOWFLAKE_WAREHOUSE": "test-warehouse",
            "SNOWFLAKE_DATABASE": "test-database",
            "SNOWFLAKE_SCHEMA": "test-schema",
            "SNOWFLAKE_ROLE": "test-role",
        }
        with patch.dict(os.environ, test_config):
            try:
                logger.info("Creating session with environment variables")
                session = Session.builder.configs(
                    {
                        "account": test_config["SNOWFLAKE_ACCOUNT"],
                        "user": test_config["SNOWFLAKE_USER"],
                        "password": test_config["SNOWFLAKE_PASSWORD"],
                        "warehouse": test_config["SNOWFLAKE_WAREHOUSE"],
                        "database": test_config["SNOWFLAKE_DATABASE"],
                        "schema": test_config["SNOWFLAKE_SCHEMA"],
                        "role": test_config["SNOWFLAKE_ROLE"],
                    }
                ).create()
                logger.info(f"Session created: {session}")
                assert session is not None
                mock_configs.assert_called_once()
            except Exception as e:
                logger.error(
                    f"Exception in test_local_environment_session_with_env_vars: {e}"
                )
                raise

    @patch("snowflake.snowpark.context.get_active_session")
    def test_local_environment_missing_env_vars(
        self, mock_get_active_session
    ):  # TODO[9c0d1e2f]: Add type annotations (see TODO.md)
        """Test that missing environment variables are handled gracefully."""
        # Mock get_active_session to fail (local environment)
        mock_get_active_session.side_effect = RuntimeError("No active session")

        # Clear any existing Snowflake environment variables
        snowflake_vars = [
            "SNOWFLAKE_ACCOUNT",
            "SNOWFLAKE_USER",
            "SNOWFLAKE_PASSWORD",
            "SNOWFLAKE_WAREHOUSE",
            "SNOWFLAKE_DATABASE",
            "SNOWFLAKE_SCHEMA",
            "SNOWFLAKE_ROLE",
        ]

        with patch.dict(os.environ, {var: "" for var in snowflake_vars}, clear=True):
            # This should fail gracefully with missing environment variables
            with pytest.raises(ValueError):
                # The function should stop execution when env vars are missing
                raise ValueError("Missing required environment variables")

    def test_environment_variable_validation(
        self,
    ):  # TODO[a0b1c2d3]: Add type annotations (see TODO.md)
        """Test that required environment variables are properly validated."""
        required_vars = [
            "SNOWFLAKE_ACCOUNT",
            "SNOWFLAKE_USER",
            "SNOWFLAKE_PASSWORD",
            "SNOWFLAKE_WAREHOUSE",
        ]

        # Test with missing required variables
        with patch.dict(os.environ, {}, clear=True):
            missing_vars = [var for var in required_vars if not os.environ.get(var)]
            assert len(missing_vars) == len(required_vars)

        # Test with all required variables present
        test_env = {
            "SNOWFLAKE_ACCOUNT": "test-account",
            "SNOWFLAKE_USER": "test-user",
            "SNOWFLAKE_PASSWORD": "test-password",
            "SNOWFLAKE_WAREHOUSE": "test-warehouse",
        }

        with patch.dict(os.environ, test_env):
            missing_vars = [var for var in required_vars if not os.environ.get(var)]
            assert len(missing_vars) == 0


class TestSessionConfiguration:
    """Test session configuration scenarios."""

    @patch("snowflake.snowpark.session.Session.SessionBuilder.configs")
    def test_session_configuration_with_all_params(
        self, mock_configs
    ):  # TODO[b1c2d3e4]: Add type annotations (see TODO.md)
        """Test session creation with all configuration parameters."""
        logger.info("Running test_session_configuration_with_all_params")
        mock_configs.return_value = Mock()
        config = {
            "account": "test-account",
            "user": "test-user",
            "password": "test-password",
            "warehouse": "test-warehouse",
            "database": "test-database",
            "schema": "test-schema",
            "role": "test-role",
        }
        try:
            session = Session.builder.configs(config).create()
            logger.info(f"Session created: {session}")
            assert session is not None
            mock_configs.assert_called_once_with(config)
        except Exception as e:
            logger.error(
                f"Exception in test_session_configuration_with_all_params: {e}"
            )
            raise

    @patch("snowflake.snowpark.session.Session.SessionBuilder.configs")
    def test_session_configuration_minimal_params(
        self, mock_configs
    ):  # TODO[c2d3e4f5]: Add type annotations (see TODO.md)
        """Test session creation with minimal required parameters."""
        logger.info("Running test_session_configuration_minimal_params")
        mock_configs.return_value = Mock()
        config = {
            "account": "test-account",
            "user": "test-user",
            "password": "test-password",
            "warehouse": "test-warehouse",
        }
        try:
            session = Session.builder.configs(config).create()
            logger.info(f"Session created: {session}")
            assert session is not None
            mock_configs.assert_called_once_with(config)
        except Exception as e:
            logger.error(f"Exception in test_session_configuration_minimal_params: {e}")
            raise


class TestAuthenticationLogic:
    """Test the authentication logic without requiring Snowflake packages."""

    def test_required_environment_variables(
        self,
    ):  # TODO[d3e4f506]: Add type annotations (see TODO.md)
        """Test that all required environment variables are identified."""
        required_vars = [
            "SNOWFLAKE_ACCOUNT",
            "SNOWFLAKE_USER",
            "SNOWFLAKE_PASSWORD",
            "SNOWFLAKE_WAREHOUSE",
        ]

        # Test validation logic
        def validate_env_vars():  # TODO[e4f50617]: Add type annotations (see TODO.md)
            missing = [var for var in required_vars if not os.environ.get(var)]
            if missing:
                raise ValueError(f"Missing required environment variables: {missing}")
            return True

        # Should fail with no environment variables
        with patch.dict(os.environ, {}, clear=True):
            with pytest.raises(ValueError):
                validate_env_vars()

        # Should succeed with all required variables
        test_env = {
            "SNOWFLAKE_ACCOUNT": "test-account",
            "SNOWFLAKE_USER": "test-user",
            "SNOWFLAKE_PASSWORD": "test-password",
            "SNOWFLAKE_WAREHOUSE": "test-warehouse",
        }

        with patch.dict(os.environ, test_env):
            assert validate_env_vars() is True

    def test_optional_environment_variables(
        self,
    ):  # TODO[f5061728]: Add type annotations (see TODO.md)
        """Test that optional environment variables are handled correctly."""
        optional_vars = ["SNOWFLAKE_DATABASE", "SNOWFLAKE_SCHEMA", "SNOWFLAKE_ROLE"]

        # Test that optional variables can be None
        with patch.dict(os.environ, {}, clear=True):
            for var in optional_vars:
                assert os.environ.get(var) is None

        # Test that optional variables can be set
        test_env = {
            "SNOWFLAKE_DATABASE": "test-db",
            "SNOWFLAKE_SCHEMA": "test-schema",
            "SNOWFLAKE_ROLE": "test-role",
        }

        with patch.dict(os.environ, test_env):
            for var, expected_value in test_env.items():
                assert os.environ.get(var) == expected_value


if __name__ == "__main__":
    pytest.main([__file__])
