requirements.txt
# Runtime Dependencies for SVG Image Generation
# Install with: pip install -r requirements.txt
# Or with uv: uv pip install -r requirements.txt

# Core Dependencies
streamlit>=1.28.0
snowflake-snowpark-python>=1.8.0
snowflake-connector-python>=3.0.0

# Optional Runtime Dependencies (uncomment as needed)
# structlog>=23.0.0
# prometheus-client>=0.17.0
python-dotenv>=1.0.0
